move `PaginatedTeams` to `types` folder
